+++ packages/did-ui-react/src/components/DeleteAccount/index.tsx
@@ -1,0 +1,10 @@
+import DeleteAccountMain, { DeleteAccountProps } from './index.component';
+import PortkeyStyleProvider from '../PortkeyStyleProvider';
+
+export default function DeleteAccount(props: DeleteAccountProps) {
+  return (
+    <PortkeyStyleProvider>
+      <DeleteAccountMain {...props} />
+    </PortkeyStyleProvider>
+  );
+}